fabtests/efa: Aggregate all fi_more tests into same file
