+++ fabtests/pytest/efa/test_rma_bw.py
@@ -90,21 +90,6 @@ def test_rma_bw_large(cmdline_args, operation_type, rma_bw_completion_semantic, 
                                completion_semantic=rma_bw_completion_semantic, message_size=67108864,
                                memory_type="host_to_host", warmup_iteration_type=0, timeout=timeout, fabric=rma_fabric)
 
-@pytest.mark.functional
-@pytest.mark.parametrize("operation_type", ["writedata", "write"])
-@pytest.mark.parametrize("iteration_type",
-                         ["5", # smaller than max batch wqe cnt (16)
-                          "48", # larger than max batch wqe cnt
-                          "128"]) # larger than window size (64)
-def test_rma_bw_use_fi_more(cmdline_args, operation_type, iteration_type, rma_bw_completion_semantic, inject_message_size, direct_rma_size, rma_fabric):
-    command = "fi_rma_bw -e rdm -j 0 --use-fi-more"
-    command = command + " -o " + operation_type
-    # rma_bw test with data verification takes longer to finish
-    timeout = max(540, cmdline_args.timeout)
-    efa_run_client_server_test(cmdline_args, command, iteration_type, rma_bw_completion_semantic,
-                               "host_to_host", direct_rma_size if rma_fabric == "efa-direct" else inject_message_size,
-                               timeout=timeout, fabric=rma_fabric)
-
 
 @pytest.mark.functional
 @pytest.mark.parametrize("comp_method", ["sread", "fd"])
